fix(blob): compile on 32-bit targets
Compare tx size as uint64 so math.MaxUint32 does not overflow int on
32-bit architectures. Add a mips cross-compile CI job to catch regressions.

+++ x/blob/ante/blob_share_decorator.go
@@ -31,7 +31,7 @@ func (d BlobShareDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool
 	}
 
 	txBytes := ctx.TxBytes()
-	if len(txBytes) > math.MaxUint32 {
+	if uint64(len(txBytes)) > math.MaxUint32 {
 		return ctx, errors.Wrapf(blobtypes.ErrBlobsTooLarge, "the tx size %d exceeds the max uint32", txBytes)
 	}
 	txSize := uint32(len(txBytes))
